+++ pkg/manager/elect/election_test.go
@@ -1,7 +1,7 @@
 // Copyright 2024 PingCAP, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package owner
+package elect
 
 import (
 	"context"
